+++ tests/api-intermediate-level/bolt.ts
@@ -1,4 +1,9 @@
 import { Keypair, type PublicKey } from "@solana/web3.js";
+import { type Position } from "../../target/types/position";
+import { type Velocity } from "../../target/types/velocity";
+import { type SystemSimpleMovement } from "../../target/types/system_simple_movement";
+import { type SystemFly } from "../../target/types/system_fly";
+import { type SystemApplyVelocity } from "../../target/types/system_apply_velocity";
 import { expect } from "chai";
 import {
   AddEntity,
@@ -12,13 +17,32 @@ import {
   RemoveAuthority,
   ApproveSystem,
   RemoveSystem,
+  type Program,
+  anchor,
   web3,
+  WORLD_PROGRAM_IDL as World,
 } from "../../clients/bolt-sdk/lib";
-import { Direction } from "../utils";
-import { Framework } from "../main";
+import { logPosition, logVelocity, Direction } from "../framework";
 
 describe("Intermediate level API", () => {
-  let framework: Framework;
+  const provider = anchor.AnchorProvider.env();
+  anchor.setProvider(provider);
+
+  const worldProgram = anchor.workspace.World as Program<World>;
+
+  const exampleComponentPosition = anchor.workspace
+    .Position as Program<Position>;
+  const exampleComponentVelocity = anchor.workspace
+    .Velocity as Program<Velocity>;
+
+  const exampleSystemSimpleMovement = (
+    anchor.workspace.SystemSimpleMovement as Program<SystemSimpleMovement>
+  ).programId;
+  const exampleSystemFly = (anchor.workspace.SystemFly as Program<SystemFly>)
+    .programId;
+  const exampleSystemApplyVelocity = (
+    anchor.workspace.SystemApplyVelocity as Program<SystemApplyVelocity>
+  ).programId;
 
   let worldPda: PublicKey;
 
@@ -35,59 +59,58 @@ describe("Intermediate level API", () => {
 
   const secondAuthority = Keypair.generate().publicKey;
 
-  it("Initialize framework", async () => {
-    framework = new Framework();
-    await framework.initialize();
-  });
-
-  it("Initialize registry", async () => {
+  it("InitializeRegistry", async () => {
     const initializeRegistry = await InitializeRegistry({
-      payer: framework.provider.wallet.publicKey,
-      connection: framework.provider.connection,
+      payer: provider.wallet.publicKey,
+      connection: provider.connection,
     });
     try {
-      await framework.provider.sendAndConfirm(initializeRegistry.transaction);
+      await provider.sendAndConfirm(initializeRegistry.transaction);
     } catch (error) {
       // This is expected to fail because the registry already exists if another api level test ran before
     }
   });
 
-  it("Initialize world", async () => {
+  it("InitializeNewWorld", async () => {
     const initializeNewWorld = await InitializeNewWorld({
-      payer: framework.provider.wallet.publicKey,
-      connection: framework.provider.connection,
+      payer: provider.wallet.publicKey,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(initializeNewWorld.transaction);
+    const signature = await provider.sendAndConfirm(
+      initializeNewWorld.transaction,
+    );
+    console.log("InitializeNewWorld signature: ", signature);
     worldPda = initializeNewWorld.worldPda; // Saved for later
   });
 
   it("Add authority", async () => {
     const addAuthority = await AddAuthority({
-      authority: framework.provider.wallet.publicKey,
-      newAuthority: framework.provider.wallet.publicKey,
+      authority: provider.wallet.publicKey,
+      newAuthority: provider.wallet.publicKey,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addAuthority.transaction, [], {
+    await provider.sendAndConfirm(addAuthority.transaction, [], {
       skipPreflight: true,
     });
-    const worldAccount = await framework.worldProgram.account.world.fetch(worldPda);
+    const worldAccount = await worldProgram.account.world.fetch(worldPda);
     expect(
       worldAccount.authorities.some((auth) =>
-        auth.equals(framework.provider.wallet.publicKey),
+        auth.equals(provider.wallet.publicKey),
       ),
     );
   });
 
   it("Add a second authority", async () => {
     const addAuthority = await AddAuthority({
-      authority: framework.provider.wallet.publicKey,
+      authority: provider.wallet.publicKey,
       newAuthority: secondAuthority,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addAuthority.transaction);
-    const worldAccount = await framework.worldProgram.account.world.fetch(worldPda);
+    const signature = await provider.sendAndConfirm(addAuthority.transaction);
+    console.log(`Add Authority signature: ${signature}`);
+    const worldAccount = await worldProgram.account.world.fetch(worldPda);
     expect(
       worldAccount.authorities.some((auth) => auth.equals(secondAuthority)),
     );
@@ -95,13 +118,14 @@ describe("Intermediate level API", () => {
 
   it("Remove an authority", async () => {
     const addAuthority = await RemoveAuthority({
-      authority: framework.provider.wallet.publicKey,
+      authority: provider.wallet.publicKey,
       authorityToDelete: secondAuthority,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addAuthority.transaction);
-    const worldAccount = await framework.worldProgram.account.world.fetch(worldPda);
+    const signature = await provider.sendAndConfirm(addAuthority.transaction);
+    console.log(`Add Authority signature: ${signature}`);
+    const worldAccount = await worldProgram.account.world.fetch(worldPda);
     expect(
       !worldAccount.authorities.some((auth) => auth.equals(secondAuthority)),
     );
@@ -109,135 +133,136 @@ describe("Intermediate level API", () => {
 
   it("InitializeNewWorld 2", async () => {
     const initializeNewWorld = await InitializeNewWorld({
-      payer: framework.provider.wallet.publicKey,
-      connection: framework.provider.connection,
+      payer: provider.wallet.publicKey,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(initializeNewWorld.transaction);
+    await provider.sendAndConfirm(initializeNewWorld.transaction);
   });
 
   it("Add entity 1", async () => {
     const addEntity = await AddEntity({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addEntity.transaction);
+    await provider.sendAndConfirm(addEntity.transaction);
     entity1Pda = addEntity.entityPda; // Saved for later
   });
 
   it("Add entity 2", async () => {
     const addEntity = await AddEntity({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addEntity.transaction);
+    await provider.sendAndConfirm(addEntity.transaction);
     entity2Pda = addEntity.entityPda; // Saved for later
   });
 
   it("Add entity 3", async () => {
     const addEntity = await AddEntity({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addEntity.transaction);
+    await provider.sendAndConfirm(addEntity.transaction);
   });
 
   it("Add entity 4 (with seed)", async () => {
     const addEntity = await AddEntity({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       world: worldPda,
       seed: Buffer.from("custom-seed"),
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addEntity.transaction);
+    await provider.sendAndConfirm(addEntity.transaction);
     entity4Pda = addEntity.entityPda;
   });
 
   it("Add entity 5", async () => {
     const addEntity = await AddEntity({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       world: worldPda,
-      connection: framework.provider.connection,
+      connection: provider.connection,
     });
-    await framework.provider.sendAndConfirm(addEntity.transaction);
+    await provider.sendAndConfirm(addEntity.transaction);
     entity5Pda = addEntity.entityPda; // Saved for later
   });
 
   it("Initialize Original Component on Entity 1, trough the world instance", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity1Pda,
-      componentId: framework.exampleComponentPosition.programId,
+      componentId: exampleComponentPosition.programId,
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
   });
 
   it("Initialize Original Component on Entity 2, trough the world instance", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity2Pda,
-      componentId: framework.exampleComponentPosition.programId,
+      componentId: exampleComponentPosition.programId,
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
   });
 
   it("Initialize Position Component on Entity 1", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity1Pda,
-      componentId: framework.exampleComponentPosition.programId,
+      componentId: exampleComponentPosition.programId,
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
     componentPositionEntity1Pda = initializeComponent.componentPda; // Saved for later
   });
 
   it("Initialize Velocity Component on Entity 1 (with seed)", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity1Pda,
-      componentId: framework.exampleComponentVelocity.programId,
+      componentId: exampleComponentVelocity.programId,
       seed: "component-velocity",
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
     componentVelocityEntity1Pda = initializeComponent.componentPda; // Saved for later
   });
 
   it("Initialize Position Component on Entity 2", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity2Pda,
-      componentId: framework.exampleComponentPosition.programId,
+      componentId: exampleComponentPosition.programId,
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
   });
 
   it("Initialize Position Component on Entity 4", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity4Pda,
-      componentId: framework.exampleComponentPosition.programId,
+      componentId: exampleComponentPosition.programId,
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
     componentPositionEntity4Pda = initializeComponent.componentPda; // Saved for later
   });
 
   it("Initialize Position Component on Entity 5 (with authority)", async () => {
     const initializeComponent = await InitializeComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity5Pda,
-      componentId: framework.exampleComponentPosition.programId,
-      authority: framework.provider.wallet.publicKey,
+      componentId: exampleComponentPosition.programId,
+      authority: provider.wallet.publicKey,
     });
-    await framework.provider.sendAndConfirm(initializeComponent.transaction);
+    await provider.sendAndConfirm(initializeComponent.transaction);
     componentPositionEntity5Pda = initializeComponent.componentPda; // Saved for later
   });
 
   it("Check Position on Entity 1 is default", async () => {
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity1Pda,
     );
+    logPosition("Default State: Entity 1", position);
     expect(position.x.toNumber()).to.equal(0);
     expect(position.y.toNumber()).to.equal(0);
     expect(position.z.toNumber()).to.equal(0);
@@ -245,28 +270,30 @@ describe("Intermediate level API", () => {
 
   it("Apply Simple Movement System (Up) on Entity 1", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemSimpleMovement.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemSimpleMovement,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
       args: {
         direction: Direction.Up,
       },
     });
-    await framework.provider.sendAndConfirm(
+    const signature = await provider.sendAndConfirm(
       applySystem.transaction,
       [],
       { skipPreflight: true },
     );
+    console.log(`Signature: ${signature}`);
 
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity1Pda,
     );
+    logPosition("Movement System: Entity 1", position);
     expect(position.x.toNumber()).to.equal(0);
     expect(position.y.toNumber()).to.equal(1);
     expect(position.z.toNumber()).to.equal(0);
@@ -274,24 +301,25 @@ describe("Intermediate level API", () => {
 
   it("Apply Simple Movement System (Right) on Entity 1", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemSimpleMovement.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemSimpleMovement,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
       args: {
         direction: Direction.Right,
       },
     });
-    await framework.provider.sendAndConfirm(applySystem.transaction);
+    await provider.sendAndConfirm(applySystem.transaction);
 
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity1Pda,
     );
+    logPosition("Movement System: Entity 1", position);
     expect(position.x.toNumber()).to.equal(1);
     expect(position.y.toNumber()).to.equal(1);
     expect(position.z.toNumber()).to.equal(0);
@@ -299,21 +327,22 @@ describe("Intermediate level API", () => {
 
   it("Apply Fly System on Entity 1", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemFly.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemFly,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
     });
-    await framework.provider.sendAndConfirm(applySystem.transaction);
+    await provider.sendAndConfirm(applySystem.transaction);
 
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity1Pda,
     );
+    logPosition("Fly System: Entity 1", position);
     expect(position.x.toNumber()).to.equal(1);
     expect(position.y.toNumber()).to.equal(1);
     expect(position.z.toNumber()).to.equal(1);
@@ -321,35 +350,37 @@ describe("Intermediate level API", () => {
 
   it("Apply System Velocity on Entity 1", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemApplyVelocity.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemApplyVelocity,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
           components: [
             {
-              componentId: framework.exampleComponentVelocity.programId,
+              componentId: exampleComponentVelocity.programId,
               seed: "component-velocity",
             },
-            { componentId: framework.exampleComponentPosition.programId },
+            { componentId: exampleComponentPosition.programId },
           ],
         },
       ],
     });
-    await framework.provider.sendAndConfirm(applySystem.transaction);
+    await provider.sendAndConfirm(applySystem.transaction);
 
-    const velocity = await framework.exampleComponentVelocity.account.velocity.fetch(
+    const velocity = await exampleComponentVelocity.account.velocity.fetch(
       componentVelocityEntity1Pda,
     );
+    logVelocity("Apply System Velocity: Entity 1", velocity);
     expect(velocity.x.toNumber()).to.equal(10);
     expect(velocity.y.toNumber()).to.equal(0);
     expect(velocity.z.toNumber()).to.equal(0);
     expect(velocity.lastApplied.toNumber()).to.not.equal(0);
 
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity1Pda,
     );
+    logPosition("Apply System Velocity: Entity 1", position);
     expect(position.x.toNumber()).to.greaterThan(1);
     expect(position.y.toNumber()).to.equal(1);
     expect(position.z.toNumber()).to.equal(1);
@@ -357,18 +388,18 @@ describe("Intermediate level API", () => {
 
   it("Apply System Velocity on Entity 1, with Clock external account", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemApplyVelocity.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemApplyVelocity,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
           components: [
             {
-              componentId: framework.exampleComponentVelocity.programId,
+              componentId: exampleComponentVelocity.programId,
               seed: "component-velocity",
             },
-            { componentId: framework.exampleComponentPosition.programId },
+            { componentId: exampleComponentPosition.programId },
           ],
         },
       ],
@@ -382,11 +413,12 @@ describe("Intermediate level API", () => {
         },
       ],
     });
-    await framework.provider.sendAndConfirm(applySystem.transaction);
+    await provider.sendAndConfirm(applySystem.transaction);
 
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity1Pda,
     );
+    logPosition("Apply System Velocity: Entity 1", position);
     expect(position.x.toNumber()).to.greaterThan(1);
     expect(position.y.toNumber()).to.equal(1);
     expect(position.z.toNumber()).to.equal(300);
@@ -394,21 +426,22 @@ describe("Intermediate level API", () => {
 
   it("Apply Fly System on Entity 4", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemFly.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemFly,
       world: worldPda,
       entities: [
         {
           entity: entity4Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
     });
-    await framework.provider.sendAndConfirm(applySystem.transaction);
+    await provider.sendAndConfirm(applySystem.transaction);
 
-    const position = await framework.exampleComponentPosition.account.position.fetch(
+    const position = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity4Pda,
     );
+    logPosition("Fly System: Entity 4", position);
     expect(position.x.toNumber()).to.equal(0);
     expect(position.y.toNumber()).to.equal(0);
     expect(position.z.toNumber()).to.equal(1);
@@ -416,7 +449,7 @@ describe("Intermediate level API", () => {
 
   it("Apply Fly System on Entity 5 (should fail with wrong authority)", async () => {
     const positionBefore =
-      await framework.exampleComponentPosition.account.position.fetch(
+      await exampleComponentPosition.account.position.fetch(
         componentPositionEntity5Pda,
       );
 
@@ -424,29 +457,30 @@ describe("Intermediate level API", () => {
 
     const applySystem = await ApplySystem({
       authority: keypair.publicKey,
-      systemId: framework.systemFly.programId,
+      systemId: exampleSystemFly,
       world: worldPda,
       entities: [
         {
           entity: entity5Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
     });
-    applySystem.transaction.recentBlockhash = (await framework.provider.connection.getLatestBlockhash()).blockhash;
-    applySystem.transaction.feePayer = framework.provider.wallet.publicKey;
+    applySystem.transaction.recentBlockhash = (await provider.connection.getLatestBlockhash()).blockhash;
+    applySystem.transaction.feePayer = provider.wallet.publicKey;
     applySystem.transaction.sign(keypair);
 
     let failed = false;
     try {
-      await framework.provider.sendAndConfirm(applySystem.transaction);
+      await provider.sendAndConfirm(applySystem.transaction);
     } catch (error) {
       failed = true;
+      // console.log("error", error);
       expect(error.logs.join("\n")).to.contain("Error Code: InvalidAuthority");
     }
     expect(failed).to.equal(true);
 
-    const positionAfter = await framework.exampleComponentPosition.account.position.fetch(
+    const positionAfter = await exampleComponentPosition.account.position.fetch(
       componentPositionEntity5Pda,
     );
 
@@ -457,91 +491,94 @@ describe("Intermediate level API", () => {
 
   it("Whitelist System", async () => {
     const approveSystem = await ApproveSystem({
-      authority: framework.provider.wallet.publicKey,
-      systemToApprove: framework.systemFly.programId,
+      authority: provider.wallet.publicKey,
+      systemToApprove: exampleSystemFly,
       world: worldPda,
     });
 
-    await framework.provider.sendAndConfirm(
+    const signature = await provider.sendAndConfirm(
       approveSystem.transaction,
       [],
       { skipPreflight: true },
     );
+    console.log(`Whitelist 2 system approval signature: ${signature}`);
 
     // Get World and check permissionless and systems
-    const worldAccount = await framework.worldProgram.account.world.fetch(worldPda);
+    const worldAccount = await worldProgram.account.world.fetch(worldPda);
     expect(worldAccount.permissionless).to.equal(false);
     expect(worldAccount.systems.length).to.be.greaterThan(0);
   });
 
   it("Whitelist System 2", async () => {
     const approveSystem = await ApproveSystem({
-      authority: framework.provider.wallet.publicKey,
-      systemToApprove: framework.systemApplyVelocity.programId,
+      authority: provider.wallet.publicKey,
+      systemToApprove: exampleSystemApplyVelocity,
       world: worldPda,
     });
 
-    await framework.provider.sendAndConfirm(
+    const signature = await provider.sendAndConfirm(
       approveSystem.transaction,
       [],
       { skipPreflight: true },
     );
+    console.log(`Whitelist 2 system approval signature: ${signature}`);
 
     // Get World and check permissionless and systems
-    const worldAccount = await framework.worldProgram.account.world.fetch(worldPda);
+    const worldAccount = await worldProgram.account.world.fetch(worldPda);
     expect(worldAccount.permissionless).to.equal(false);
     expect(worldAccount.systems.length).to.be.greaterThan(0);
   });
 
   it("Apply Fly System on Entity 1", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemFly.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemFly,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
     });
-    await framework.provider.sendAndConfirm(applySystem.transaction);
+    await provider.sendAndConfirm(applySystem.transaction);
   });
 
   it("Remove System 1", async () => {
     const approveSystem = await RemoveSystem({
-      authority: framework.provider.wallet.publicKey,
-      systemToRemove: framework.systemFly.programId,
+      authority: provider.wallet.publicKey,
+      systemToRemove: exampleSystemFly,
       world: worldPda,
     });
 
-    await framework.provider.sendAndConfirm(
+    const signature = await provider.sendAndConfirm(
       approveSystem.transaction,
       [],
       { skipPreflight: true },
     );
+    console.log(`Whitelist 2 system approval signature: ${signature}`);
 
     // Get World and check permissionless and systems
-    const worldAccount = await framework.worldProgram.account.world.fetch(worldPda);
+    const worldAccount = await worldProgram.account.world.fetch(worldPda);
     expect(worldAccount.permissionless).to.equal(false);
     expect(worldAccount.systems.length).to.be.greaterThan(0);
   });
 
   it("Apply Invalid Fly System on Entity 1", async () => {
     const applySystem = await ApplySystem({
-      authority: framework.provider.wallet.publicKey,
-      systemId: framework.systemFly.programId,
+      authority: provider.wallet.publicKey,
+      systemId: exampleSystemFly,
       world: worldPda,
       entities: [
         {
           entity: entity1Pda,
-          components: [{ componentId: framework.exampleComponentPosition.programId }],
+          components: [{ componentId: exampleComponentPosition.programId }],
         },
       ],
     });
     let invalid = false;
     try {
-      await framework.provider.sendAndConfirm(applySystem.transaction);
+      await provider.sendAndConfirm(applySystem.transaction);
     } catch (error) {
       expect(error.logs.join(" ")).to.contain("Error Code: SystemNotApproved");
       invalid = true;
@@ -552,16 +589,17 @@ describe("Intermediate level API", () => {
   it("Check invalid component init without CPI", async () => {
     let invalid = false;
     try {
-      await framework.exampleComponentPosition.methods
+      await exampleComponentPosition.methods
         .initialize()
         .accounts({
-          payer: framework.provider.wallet.publicKey,
+          payer: provider.wallet.publicKey,
           data: componentPositionEntity5Pda,
           entity: entity5Pda,
-          authority: framework.provider.wallet.publicKey,
+          authority: provider.wallet.publicKey,
         })
         .rpc();
     } catch (error) {
+      // console.log("error", error);
       expect(error.message).to.contain("Error Code: InvalidCaller");
       invalid = true;
     }
@@ -571,11 +609,11 @@ describe("Intermediate level API", () => {
   it("Check invalid component update without CPI", async () => {
     let invalid = false;
     try {
-      await framework.exampleComponentPosition.methods
+      await exampleComponentPosition.methods
         .update(Buffer.from(""))
         .accounts({
           boltComponent: componentPositionEntity4Pda,
-          authority: framework.provider.wallet.publicKey,
+          authority: provider.wallet.publicKey,
           sessionToken: null
         })
         .rpc();
@@ -588,17 +626,17 @@ describe("Intermediate level API", () => {
 
   it("Check component delegation", async () => {
     const delegateComponent = await DelegateComponent({
-      payer: framework.provider.wallet.publicKey,
+      payer: provider.wallet.publicKey,
       entity: entity1Pda,
-      componentId: framework.exampleComponentPosition.programId,
+      componentId: exampleComponentPosition.programId,
     });
 
-    await framework.provider.sendAndConfirm(
+    const txSign = await provider.sendAndConfirm(
       delegateComponent.transaction,
       [],
       { skipPreflight: true, commitment: "confirmed" },
     );
-    const acc = await framework.provider.connection.getAccountInfo(
+    const acc = await provider.connection.getAccountInfo(
       delegateComponent.componentPda,
     );
     expect(acc?.owner.toBase58()).to.equal(DELEGATION_PROGRAM_ID.toBase58());
